refactor: add `typing.py`
